# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Books Service
    description: Service for managing Books with CRUD operations.
    version: ""
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /book:
        post:
            tags:
                - books
            summary: create-book books
            description: Create a single book.
            operationId: books#create-book
            parameters:
                - name: Create-BookRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/BooksCreateBookRequestBody'
                    required:
                        - title
                        - author
                        - book_cover
                        - published_at
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/BooksCreateBookResponseBody'
            schemes:
                - http
        patch:
            tags:
                - books
            summary: update-book books
            description: Updates a book by the given id.
            operationId: books#update-book
            parameters:
                - name: Update-BookRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/BooksUpdateBookRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/BooksUpdateBookResponseBody'
            schemes:
                - http
    /book/{id}:
        get:
            tags:
                - books
            summary: get-book books
            description: Retrieve a book by id.
            operationId: books#get-book
            parameters:
                - name: id
                  in: path
                  description: Book id
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/BooksGetBookResponseBody'
            schemes:
                - http
        delete:
            tags:
                - books
            summary: delete-book books
            description: Delete a single book.
            operationId: books#delete-book
            parameters:
                - name: id
                  in: path
                  description: Book id
                  required: true
                  type: integer
            responses:
                "204":
                    description: No Content response.
            schemes:
                - http
    /books/{skip}/{take}:
        get:
            tags:
                - books
            summary: get-books books
            description: Get paginated books by specifying the number of books to skip and take.
            operationId: books#get-books
            parameters:
                - name: skip
                  in: path
                  description: Number of books to skip
                  required: true
                  type: integer
                - name: take
                  in: path
                  description: Number of books to take after skip
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/BooksGetBooksResponseBody'
            schemes:
                - http
    /openapi.json:
        get:
            tags:
                - books
            summary: Download ./gen/http/openapi.json
            operationId: books#/openapi.json
            responses:
                "200":
                    description: File downloaded
                    schema:
                        type: file
            schemes:
                - http
definitions:
    BooksCreateBookRequestBody:
        title: BooksCreateBookRequestBody
        type: object
        properties:
            author:
                type: string
                description: The author of the book.
                example: rpd
                minLength: 3
                maxLength: 50
            book_cover:
                type: string
                description: The cover image of the book.
                example: gwz
                minLength: 15
            published_at:
                type: string
                description: The date at which the book was published.
                example: "2007-12-06"
                format: date
            title:
                type: string
                description: The title of the book.
                example: "4"
                minLength: 1
                maxLength: 100
        example:
            author: 2kt
            book_cover: s4i
            published_at: "1984-02-22"
            title: "68"
        required:
            - title
            - author
            - book_cover
            - published_at
    BooksCreateBookResponseBody:
        title: 'Mediatype identifier: createbookresponse; view=default'
        type: object
        properties:
            author:
                type: string
                description: The author of the book.
                example: Possimus impedit.
            book_cover:
                type: string
                description: The cover image of the book.
                example: Tempora labore quam maiores ut nihil.
            id:
                type: integer
                description: The unique id of the book.
                example: 318504118629116124
                format: int64
            published_at:
                type: string
                description: The date at which the book was published.
                example: Consequatur est voluptas fugit id.
            title:
                type: string
                description: The title of the book.
                example: Id nihil.
        description: Create-BookResponseBody result type (default view)
        example:
            author: Aliquam omnis.
            book_cover: Dolorem enim porro qui.
            id: 8852360738290317152
            published_at: Ab pariatur.
            title: Fugit ab error quisquam explicabo sed minima.
        required:
            - id
            - title
            - book_cover
            - author
            - published_at
    BooksGetBookResponseBody:
        title: 'Mediatype identifier: getbookresponse; view=default'
        type: object
        properties:
            author:
                type: string
                description: The author of the book.
                example: Autem nemo.
            book_cover:
                type: string
                description: The cover image of the book.
                example: Sed aperiam.
            id:
                type: integer
                description: The unique id of the book.
                example: 8391061508749378211
                format: int64
            published_at:
                type: string
                description: The date at which the book was published.
                example: Quia accusamus dolorum omnis.
            title:
                type: string
                description: The title of the book.
                example: Ipsam aperiam itaque.
        description: Get-BookResponseBody result type (default view)
        example:
            author: Consequatur ipsam at quia.
            book_cover: A aut asperiores.
            id: 4258842219939479309
            published_at: Unde et ducimus magnam nemo porro aut.
            title: Recusandae reiciendis.
        required:
            - id
            - title
            - book_cover
            - author
            - published_at
    BooksGetBooksResponseBody:
        title: 'Mediatype identifier: getbooksresponse; view=default'
        type: object
        properties:
            books:
                $ref: '#/definitions/ResultbookResponseBodyCollection'
        description: Get-BooksResponseBody result type (default view)
        example:
            books:
                - author: Assumenda recusandae corrupti.
                  book_cover: Doloribus similique possimus quia.
                  id: 3560148692509608818
                  published_at: Possimus eos facere placeat.
                  title: Vitae a amet distinctio reprehenderit ea inventore.
                - author: Assumenda recusandae corrupti.
                  book_cover: Doloribus similique possimus quia.
                  id: 3560148692509608818
                  published_at: Possimus eos facere placeat.
                  title: Vitae a amet distinctio reprehenderit ea inventore.
                - author: Assumenda recusandae corrupti.
                  book_cover: Doloribus similique possimus quia.
                  id: 3560148692509608818
                  published_at: Possimus eos facere placeat.
                  title: Vitae a amet distinctio reprehenderit ea inventore.
                - author: Assumenda recusandae corrupti.
                  book_cover: Doloribus similique possimus quia.
                  id: 3560148692509608818
                  published_at: Possimus eos facere placeat.
                  title: Vitae a amet distinctio reprehenderit ea inventore.
    BooksUpdateBookRequestBody:
        title: BooksUpdateBookRequestBody
        type: object
        properties:
            author:
                type: string
                description: The author of the book.
                example: 5kj
                minLength: 3
                maxLength: 50
            book_cover:
                type: string
                description: The cover image of the book.
                example: al5
                minLength: 15
            id:
                type: integer
                description: The unique id of the book.
                example: 6149502838993478485
                minimum: 1
            published_at:
                type: string
                description: The date at which the book was published.
                example: "1988-05-03"
                format: date
            title:
                type: string
                description: The title of the book.
                example: pw
                minLength: 1
                maxLength: 100
        example:
            author: r8f
            book_cover: 77e
            id: 3928243132458917846
            published_at: "1972-12-02"
            title: fj
    BooksUpdateBookResponseBody:
        title: 'Mediatype identifier: updatebookresponse; view=default'
        type: object
        properties:
            author:
                type: string
                description: The author of the book.
                example: Quos velit deleniti.
            book_cover:
                type: string
                description: The cover image of the book.
                example: Quia consequatur doloremque voluptatem.
            id:
                type: integer
                description: The unique id of the book.
                example: 6037058422783492289
                format: int64
            published_at:
                type: string
                description: The date at which the book was published.
                example: Est nisi ut tempore.
            title:
                type: string
                description: The title of the book.
                example: Molestias omnis minus quia aut vitae et.
        description: Update-BookResponseBody result type (default view)
        example:
            author: Nobis deleniti.
            book_cover: Expedita omnis.
            id: 5429017585288396179
            published_at: Ad enim voluptatem dolorum ut et harum.
            title: Voluptate et eos eius qui.
        required:
            - id
            - title
            - book_cover
            - author
            - published_at
    ResultbookResponseBody:
        title: 'Mediatype identifier: resultbook; view=default'
        type: object
        properties:
            author:
                type: string
                description: The author of the book.
                example: Ad ipsum quia doloribus.
            book_cover:
                type: string
                description: The cover image of the book.
                example: Animi doloremque occaecati natus quo facilis eum.
            id:
                type: integer
                description: The unique id of the book.
                example: 9141884749962453202
                format: int64
            published_at:
                type: string
                description: The date at which the book was published.
                example: Iusto autem praesentium aperiam.
            title:
                type: string
                description: The title of the book.
                example: Omnis nemo tempora sed et.
        description: ResultbookResponseBody result type (default view)
        example:
            author: Libero sunt sit voluptas.
            book_cover: Praesentium dicta vel architecto officiis.
            id: 2360347935652067765
            published_at: Alias quis.
            title: Voluptas consequatur maiores ut ad dolor ut.
        required:
            - id
            - title
            - book_cover
            - author
            - published_at
    ResultbookResponseBodyCollection:
        title: 'Mediatype identifier: resultbook; type=collection; view=default'
        type: array
        items:
            $ref: '#/definitions/ResultbookResponseBody'
        description: ResultbookCollectionResponseBody is the result type for an array of ResultbookResponseBody (default view)
        example:
            - author: Assumenda recusandae corrupti.
              book_cover: Doloribus similique possimus quia.
              id: 3560148692509608818
              published_at: Possimus eos facere placeat.
              title: Vitae a amet distinctio reprehenderit ea inventore.
            - author: Assumenda recusandae corrupti.
              book_cover: Doloribus similique possimus quia.
              id: 3560148692509608818
              published_at: Possimus eos facere placeat.
              title: Vitae a amet distinctio reprehenderit ea inventore.